+++ src/include/platform.hpp
@@ -5,7 +5,7 @@
 #include <cstddef>
 
 #ifdef _WIN32
-#define WIN32_LEAN_AND_MEAN
+#include "platform_win98.h"
 #include <windows.h>
 #endif